Rename objects/environment.rs to ground.rs, delete unused mesh_utils.rs
Prep for the bevy-free module seam (ADR-0001): mesh_utils.rs was never
declared as a module in lib.rs and had zero references anywhere in the
repo, so it's dead code. environment.rs renamed to ground.rs to match
its GroundPlane content.
